+++ packages/docs/src/helpers/colors.ts
@@ -1,5 +1,5 @@
 import { paramCase } from 'change-case';
-import * as tokens from '@sick-design-system/design-tokens';
+import * as tokens from '@sick-design-system/tokens';
 
 /**
  * Get all colors from a palette as object