+++ packages/tokens/scripts/build.js
@@ -1,19 +1,21 @@
 import { TokenBuilder } from './build-tokens.js';
 import { addMissingTokens } from './add-missing-tokens.js';
-import 'dotenv/config';
 
-export const createThemes = async (config) => {
+// TODO: Remove LibraryPrefix hardcoded
+const createThemes = async () => {
   const tokenBuilder = new TokenBuilder({
     sourcePaths: [
-      '../src/colors/figma-tokens/primitives.json',
-      '../src/colors/figma-tokens/globals.json',
-      '../src/colors/figma-tokens/semantics/*.json',
+      '../src/figma-tokens/colors/primitives.json',
+      '../src/figma-tokens/globals.json',
+      '../src/figma-tokens/semantics/*.json',
     ],
     buildPath: '../dist/',
-    prefix: `${config.libraryPrefix}-`,
+    prefix: 'sds-',
   });
   tokenBuilder.buildTokens();
 
-  await addMissingTokens(config.libraryPrefix);
+  await addMissingTokens('sds');
   return Promise.resolve(true);
 };
+
+createThemes();