+++ packages/docs/CHANGELOG.md
@@ -21,7 +21,7 @@ The following icons are used in commit messages and this changelog.
 ### 📚 Doc
 
 - #22: Added Welcome page and colors
-- #22: Added `@sick-design-system/design-tokens` Changelog and Readme, as well as color tokens
+- #22: Added `@sick-design-system/tokens` Changelog and Readme, as well as color tokens
 
 ### ✨ Features